+++ tests/test_arm.c
@@ -187,7 +187,7 @@ static void test()
 		},
 	};
 
-	size_t address = 0x1000;
+	uint64_t address = 0x1000;
 	cs_insn *insn;
 	int i;
